+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Logic gates".
 - Tests for "Master of mayhem".
 - Tests for "Body weight is a girl's secret".
+- Tests for "Pirate's treasure".
 
 ## [3.8.0] - 2022-04-30
 ### Added